fixed return type of reset

+++ gym_maze/envs/maze_env.py
@@ -91,7 +91,7 @@ class MazeEnv(gym.Env):
 
     def reset(self):
         self.maze_view.reset_robot()
-        self.state = np.zeros(2)
+        self.state = np.zeros(2, dtype=np.int64)
         self.steps_beyond_done = None
         self.done = False
         return self.state
